Final Report/Team10/NestWise-main/backend-userFinancedata/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
pydantic
python-dateutil
celery[redis]
boto3
python-multipart
alembic
pydantic-settings
python-jose